[message] rename kubernetes to k8s_event
